+++ tensorboard/webapp/notification_center/_data_source/BUILD
@@ -29,3 +29,18 @@ tf_ts_library(
         "@npm//rxjs",
     ],
 )
+
+tf_ts_library(
+    name = "testing",
+    testonly = True,
+    srcs = [
+        "testing.ts",
+    ],
+    deps = [
+        ":_data_source",
+        ":types",
+        "//tensorboard/webapp/notification_center/_redux:types",
+        "@npm//@angular/core",
+        "@npm//rxjs",
+    ],
+)
