+++ tensorboard/webapp/notification_center/_redux/notification_center_effects_test.ts
@@ -31,9 +31,9 @@ import {
 } from './notification_center_effects';
 import {CategoryEnum} from './notification_center_types';
 import {
-  createNotification,
+  buildNotificationResponse,
   provideTestingNotificationCenterDataSource,
-} from './testing';
+} from '../_data_source/testing';
 
 describe('notification center effects', () => {
   let dataSource: NotificationCenterDataSource;
@@ -76,7 +76,7 @@ describe('notification center effects', () => {
   it('fetches notifications on initial load', () => {
     actions$.next(TEST_ONLY.initAction());
     fetchNotificationsSpy.and.returnValue(fetchNotificationSubject);
-    fetchNotificationSubject.next(createNotification());
+    fetchNotificationSubject.next(buildNotificationResponse());
 
     expect(fetchNotificationsSpy).toHaveBeenCalled();
     expect(actualActions).toEqual([
