+++ tensorboard/webapp/notification_center/_redux/testing.ts
@@ -12,20 +12,11 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 ==============================================================================*/
-import {Injectable} from '@angular/core';
-import {of} from 'rxjs';
 import {
-  NotificationCenterDataSource,
-  NotificationCenterResponse as DataSourceNotifications,
-} from '../_data_source';
-import {
-  CategoryEnum,
   NotificationState,
   NOTIFICATION_FEATURE_KEY,
   State,
 } from './notification_center_types';
-
-/** @typehack */ import * as _typeHackRxjs from 'rxjs';
 
 export function buildNotificationState(
   override: Partial<NotificationState>
@@ -41,37 +32,4 @@ export function buildStateFromNotificationState(
   runsState: NotificationState
 ): State {
   return {[NOTIFICATION_FEATURE_KEY]: runsState};
-}
-
-@Injectable()
-export class TestingNotificationCenterDataSource
-  implements NotificationCenterDataSource {
-  fetchNotifications() {
-    return of({
-      notifications: [],
-    });
-  }
-}
-
-export function provideTestingNotificationCenterDataSource() {
-  return [
-    TestingNotificationCenterDataSource,
-    {
-      provide: NotificationCenterDataSource,
-      useExisting: TestingNotificationCenterDataSource,
-    },
-  ];
-}
-
-export function createNotification(): DataSourceNotifications {
-  return {
-    notifications: [
-      {
-        category: CategoryEnum.WHATS_NEW,
-        dateInMs: 123,
-        title: 'test title',
-        content: 'random content',
-      },
-    ],
-  };
 }
